@@ -14,9 +14,14 @@ namespace HealthChecks.UI.Sample
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            app.UseHealthChecksUI();
+            app
+                .UseRouting()
+                .UseEndpoints(config =>
+                {
+                    config.MapHealthChecksUI();
+                });
         }
     }
 }
