@@ -6,7 +6,11 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Polly;
+using Polly.Extensions.Http;
+using Polly.Timeout;
 using System;
+using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -25,7 +29,7 @@ namespace HealthChecks.UIAndApi
         {
             //
             //  This project configure health checks for asp.net core project and UI
-            //  in the same project. Typically health checks and UI are on different projects 
+            //  in the same project. Typically health checks and UI are on different projects
             //  UI exist also as container image
             //
 
@@ -43,13 +47,12 @@ namespace HealthChecks.UIAndApi
                 //        .CheckPod("myapp-pod", p =>  p.Metadata.Labels["app"] == "myapp" );
                 //})
                 .Services
-                .AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                .AddControllers();
 
             //
             //   below show howto use default policy handlers ( polly )
             //   with httpclient on asp.net core also
-            //   on uri health checks 
+            //   on uri health checks
             //
 
             //var retryPolicy = HttpPolicyExtensions
@@ -58,19 +61,28 @@ namespace HealthChecks.UIAndApi
             //    .RetryAsync(5);
 
             //services.AddHttpClient("uri-group") //default healthcheck registration name for uri ( you can change it on AddUrlGroup )
-            //    .AddPolicyHandler(retryPolicy);
+            //    .AddPolicyHandler(retryPolicy)
+            //    .ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler()
+            //    {
+            //        ClientCertificateOptions = ClientCertificateOption.Manual,
+            //        ServerCertificateCustomValidationCallback = (httpRequestMessage, cert, cetChain, policyErrors) =>
+            //        {
+            //            return true;
+            //        }
+            //    });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-           app.UseHealthChecks("/healthz",new HealthCheckOptions
-           {
-               Predicate = _=> true,
-               ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse
-           })
-           .UseHealthChecksUI()
-           .UseMvc();
+            app
+                .UseRouting()
+                .UseEndpoints(config =>
+                {
+                    config.MapHealthChecksUI();
+                    config.MapDefaultControllerRoute();
+                });
+
         }
     }
 
