@@ -19,33 +19,40 @@ namespace HealthChecks.UIAndApi
 
             services
                 .AddHealthChecksUI()
-//                .AddHealthChecksUI(setupSettings: settings =>
-//                {
-//                    settings
-//                        .AddHealthCheckEndpoint("api1", "http://localhost:8001/custom/healthz")
-//                        .AddWebhookNotification("webhook1", "http://webhook", "mypayload")
-//                        .SetEvaluationTimeInSeconds(16);
-//                })
+                //                .AddHealthChecksUI(setupSettings: settings =>
+                //                {
+                //                    settings
+                //                        .AddHealthCheckEndpoint("api1", "http://localhost:8001/custom/healthz")
+                //                        .AddWebhookNotification("webhook1", "http://webhook", "mypayload")
+                //                        .SetEvaluationTimeInSeconds(16);
+                //                })
                 .AddHealthChecks()
                 .AddUrlGroup(new Uri("http://httpbin.org/status/200"), name: "uri-1")
                 .AddUrlGroup(new Uri("http://httpbin.org/status/200"), name: "uri-2")
                 .AddUrlGroup(new Uri("http://httpbin.org/status/500"), name: "uri-3")
                 .Services
-                .AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                .AddControllers();
         }
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            app.UseHealthChecks("/healthz", new HealthCheckOptions
-            {
-                Predicate = _ => true,
-                ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse
-            })
-            .UseHealthChecksUI(setup =>
-            {
-                setup.UIPath = "/show-health-ui"; // this is ui path in your browser
-                setup.ApiPath = "/health-ui-api"; // the UI ( spa app )  use this path to get information from the store ( this is NOT the healthz path, is internal ui api )
-            }).UseMvc();
+            app.UseRouting()
+               .UseEndpoints(config =>
+                {
+                    config.MapHealthChecks("/healthz", new HealthCheckOptions
+                    {
+                        Predicate = _ => true,
+                        ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse
+                    });
+
+                    config.MapHealthChecksUI(setup =>
+                    {
+                        setup.UIPath = "/show-health-ui"; // this is ui path in your browser
+                        setup.ApiPath = "/health-ui-api"; // the UI ( spa app )  use this path to get information from the store ( this is NOT the healthz path, is internal ui api )
+                    });
+
+                    config.MapDefaultControllerRoute();
+                });
         }
     }
 }
