@@ -1,39 +1,61 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using Newtonsoft.Json.Serialization;
+using System;
 using System.Threading.Tasks;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace HealthChecks.UI.Client
 {
     public static class UIResponseWriter
     {
         const string DEFAULT_CONTENT_TYPE = "application/json";
-        public static Task WriteHealthCheckUIResponse(HttpContext httpContext, HealthReport report)
+
+        public static Task WriteHealthCheckUIResponse(HttpContext httpContext, HealthReport report) => WriteHealthCheckUIResponse(httpContext, report, null);
+
+        public static Task WriteHealthCheckUIResponse(HttpContext httpContext, HealthReport report, Action<JsonSerializerOptions> jsonConfigurator)
         {
             var response = "{}";
 
             if (report != null)
             {
-                var settings = new JsonSerializerSettings()
+                var settings = new JsonSerializerOptions()
                 {
-                    ContractResolver = new CamelCasePropertyNamesContractResolver(),
-                    NullValueHandling = NullValueHandling.Ignore,
-                    ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
+                    AllowTrailingCommas = true,
+                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                    IgnoreNullValues = true,
                 };
 
-                settings.Converters.Add(new StringEnumConverter());
+                jsonConfigurator?.Invoke(settings);
+                
+                settings.Converters.Add(new JsonStringEnumConverter());
+
+                //for compatibility with older UI versions ( <3.0 ) we arrange
+                //timepan serialization as s
+                settings.Converters.Add(new TimeSpanConverter());
 
                 httpContext.Response.ContentType = DEFAULT_CONTENT_TYPE;
 
                 var uiReport = UIHealthReport
                     .CreateFrom(report);
 
-                response = JsonConvert.SerializeObject(uiReport, settings);
+                response = JsonSerializer.Serialize(uiReport, settings);
             }
 
             return httpContext.Response.WriteAsync(response);
+        }
+    }
+
+    internal class TimeSpanConverter
+        : JsonConverter<TimeSpan>
+    {
+        public override TimeSpan Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        {
+            return default;
+        }
+        public override void Write(Utf8JsonWriter writer, TimeSpan value, JsonSerializerOptions options)
+        {
+            writer.WriteStringValue(value.ToString());
         }
     }
 }
