@@ -20,9 +20,10 @@ namespace Microsoft.Extensions.DependencyInjection
         /// the default status of <see cref="HealthStatus.Unhealthy"/> will be reported.
         /// </param>
         /// <param name="tags">A list of tags that can be used to filter sets of health checks. Optional.</param>
+        /// <param name="timeout">An optional System.TimeSpan representing the timeout of the check.</param>
         /// <returns>The <see cref="IHealthChecksBuilder"/>.</returns></param>
         public static IHealthChecksBuilder AddAzureKeyVault(this IHealthChecksBuilder builder, Action<AzureKeyVaultOptions> setup,
-            string name = default, HealthStatus? failureStatus = default, IEnumerable<string> tags = default)
+            string name = default, HealthStatus? failureStatus = default, IEnumerable<string> tags = default, TimeSpan? timeout = default)
         {
             var options = new AzureKeyVaultOptions();
             setup?.Invoke(options);
@@ -31,7 +32,8 @@ namespace Microsoft.Extensions.DependencyInjection
                name ?? KEYVAULT_NAME,
                sp => new AzureKeyVaultHealthCheck(options),
                failureStatus,
-               tags));
+               tags,
+               timeout));
         }
     }
 }
