@@ -12,16 +12,19 @@ namespace HealthChecks.AzureServiceBus
         : IHealthCheck
     {
         private const string TEST_MESSAGE = "HealthCheckTestMessage";
-        private readonly string _connectionString;
-        private readonly string _topicName;
         private static readonly ConcurrentDictionary<string, TopicClient> _topicClient = new ConcurrentDictionary<string, TopicClient>();
 
-        public AzureServiceBusTopicHealthCheck(string connectionString, string topicName)
+        private readonly string _connectionString;
+        private readonly string _topicName;
+        private readonly Action<Message> _configuringMessage;
+
+        public AzureServiceBusTopicHealthCheck(string connectionString, string topicName, Action<Message> configuringMessage = null)
         {
             if (string.IsNullOrEmpty(connectionString))
             {
                 throw new ArgumentNullException(nameof(connectionString));
             }
+
             if (string.IsNullOrEmpty(topicName))
             {
                 throw new ArgumentNullException(nameof(topicName));
@@ -29,6 +32,7 @@ namespace HealthChecks.AzureServiceBus
 
             _connectionString = connectionString;
             _topicName = topicName;
+            _configuringMessage = configuringMessage;
         }
         public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
         {
@@ -45,8 +49,11 @@ namespace HealthChecks.AzureServiceBus
                     }
                 }
 
-                var scheduledMessageId = await topicClient.ScheduleMessageAsync(
-                    new Message(Encoding.UTF8.GetBytes(TEST_MESSAGE)),
+                var message = new Message(Encoding.UTF8.GetBytes(TEST_MESSAGE));
+                
+                _configuringMessage?.Invoke(message);
+
+                var scheduledMessageId = await topicClient.ScheduleMessageAsync(message, 
                     new DateTimeOffset(DateTime.UtcNow).AddHours(2));
 
                 await topicClient.CancelScheduledMessageAsync(scheduledMessageId);
