@@ -1,5 +1,6 @@
 using HealthChecks.EventStore;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+using System;
 using System.Collections.Generic;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -20,14 +21,16 @@ namespace Microsoft.Extensions.DependencyInjection
         /// the default status of <see cref="HealthStatus.Unhealthy"/> will be reported.
         /// </param>
         /// <param name="tags">A list of tags that can be used to filter sets of health checks. Optional.</param>
+        /// <param name="timeout">An optional System.TimeSpan representing the timeout of the check.</param>
         /// <returns>The <see cref="IHealthChecksBuilder"/>.</returns>
-        public static IHealthChecksBuilder AddEventStore(this IHealthChecksBuilder builder, string eventStoreConnection, string login = default, string password = default, string name = default, HealthStatus? failureStatus = default, IEnumerable<string> tags = default)
+        public static IHealthChecksBuilder AddEventStore(this IHealthChecksBuilder builder, string eventStoreConnection, string login = default, string password = default, string name = default, HealthStatus? failureStatus = default, IEnumerable<string> tags = default, TimeSpan? timeout = default)
         {
             return builder.Add(new HealthCheckRegistration(
                 name ?? NAME,
                 sp => new EventStoreHealthCheck(eventStoreConnection, login, password),
                 failureStatus,
-                tags));
+                tags,
+                timeout));
         }
     }
 }
