@@ -1,5 +1,6 @@
 ﻿using HealthChecks.Oracle;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+using System;
 using System.Collections.Generic;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -16,17 +17,19 @@ namespace Microsoft.Extensions.DependencyInjection
     /// the default status of <see cref="HealthStatus.Unhealthy"/> will be reported.
     /// </param>
     /// <param name="tags">A list of tags that can be used to filter sets of health checks. Optional.</param>
+    /// <param name="timeout">An optional System.TimeSpan representing the timeout of the check.</param>
     /// <returns>The <see cref="IHealthChecksBuilder"/>.</returns></param>
     public static class OracleHealthCheckBuilderExtensions
     {
         const string NAME = "oracle";
-        public static IHealthChecksBuilder AddOracle(this IHealthChecksBuilder builder, string connectionString, string healthQuery = "select * from v$version", string name = default, HealthStatus? failureStatus = default, IEnumerable<string> tags = default)
+        public static IHealthChecksBuilder AddOracle(this IHealthChecksBuilder builder, string connectionString, string healthQuery = "select * from v$version", string name = default, HealthStatus? failureStatus = default, IEnumerable<string> tags = default,TimeSpan? timeout = default)
         { 
             return builder.Add(new HealthCheckRegistration(
                 name ?? NAME,
                 sp => new OracleHealthCheck(connectionString, healthQuery),
                 failureStatus,
-                tags));
+                tags,
+                timeout));
         }
     }
 }
