@@ -22,15 +22,13 @@ namespace Microsoft.Extensions.DependencyInjection
             string endpoint, string job, string instance = null)
         {
             builder.Services
-                .AddSingleton<IHealthCheckPublisher>(sp => new PrometheusGatewayPublisher(endpoint, job, instance));
+                .AddHttpClient();
 
-            return builder;
-        }
-        internal static IHealthChecksBuilder AddPrometheusGatewayPublisher(this IHealthChecksBuilder builder,
-            HttpClient client, string endpoint, string job, string instance = null)
-        {
             builder.Services
-                .AddSingleton<IHealthCheckPublisher>(sp => new PrometheusGatewayPublisher(client, endpoint, job, instance));
+                .AddSingleton<IHealthCheckPublisher>(sp =>
+                {
+                    return new PrometheusGatewayPublisher(() => sp.GetRequiredService<IHttpClientFactory>().CreateClient(), endpoint, job, instance);
+                });
 
             return builder;
         }
