@@ -8,44 +8,7 @@ namespace Microsoft.Extensions.DependencyInjection
     public static class RavenDBHealthCheckBuilderExtensions
     {
         const string NAME = "ravendb";
-        /// <summary>
-        /// </summary>
-        /// <param name="builder">The <see cref="IHealthChecksBuilder"/>.</param>
-        /// <param name="connectionString">Connection string to RavenDB.</param>
-        /// <param name="databaseName">The specified database to check.</param>
-        /// <param name="name">
-        /// The health check name. Optional. If <see langword="null"/> the type name 'ravendb' will be used for the name.
-        /// </param>
-        /// <param name="failureStatus">
-        /// The <see cref="HealthStatus"/> that should be reported when the health check fails. Optional. If <see langword="null"/> then
-        /// the default status of <see cref="HealthStatus.Unhealthy"/> will be reported.
-        /// </param>
-        /// <param name="tags">A list of tags that can be used to filter sets of health checks. Optional.</param>
-        /// <returns>The <see cref="IHealthChecksBuilder"/>.</returns>
-        [Obsolete("This method is deprecated. Use RavenDBOptions overload.")]
-        public static IHealthChecksBuilder AddRavenDB(
-            this IHealthChecksBuilder builder,
-            string connectionString,
-            string databaseName = default,
-            string name = default,
-            HealthStatus? failureStatus = default,
-            IEnumerable<string> tags = default)
-        {
-            if (connectionString == null)
-            {
-                throw new ArgumentNullException(nameof(connectionString));
-            }
-
-            return builder.AddRavenDB(
-                _ =>
-                {
-                    _.Urls = new[] {connectionString};
-                    _.Database = databaseName;
-                },
-                name,
-                failureStatus,
-                tags);
-        }
+       
         /// <summary>
         /// Add a health check for RavenDB.
         /// </summary>
@@ -59,13 +22,15 @@ namespace Microsoft.Extensions.DependencyInjection
         /// the default status of <see cref="HealthStatus.Unhealthy"/> will be reported.
         /// </param>
         /// <param name="tags">A list of tags that can be used to filter sets of health checks. Optional.</param>
+        /// <param name="timeout">An optional System.TimeSpan representing the timeout of the check.</param>
         /// <returns>The <see cref="IHealthChecksBuilder"/>.</returns>
         public static IHealthChecksBuilder AddRavenDB(
             this IHealthChecksBuilder builder,
             Action<RavenDBOptions> setup,
             string name = default,
             HealthStatus? failureStatus = default,
-            IEnumerable<string> tags = default)
+            IEnumerable<string> tags = default,
+            TimeSpan? timeout = default)
         {
             var options = new RavenDBOptions();
             setup?.Invoke(options);
