@@ -17,13 +17,14 @@ namespace HealthChecks.Redis
         {
             _redisConnectionString = redisConnectionString ?? throw new ArgumentNullException(nameof(redisConnectionString));
         }
-        public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
+
+        public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
         {
             try
             {
                 if (!_connections.TryGetValue(_redisConnectionString, out ConnectionMultiplexer connection))
                 {
-                    connection = ConnectionMultiplexer.Connect(_redisConnectionString);
+                    connection = await ConnectionMultiplexer.ConnectAsync(_redisConnectionString);
 
                     if (!_connections.TryAdd(_redisConnectionString, connection))
                     {
@@ -33,17 +34,15 @@ namespace HealthChecks.Redis
                     }
                 }
 
-                connection.GetDatabase()
-                    .Ping();
+                await connection.GetDatabase()
+                    .PingAsync();
 
-                return Task.FromResult(
-                    HealthCheckResult.Healthy());
-            }
+				return HealthCheckResult.Healthy();
+			}
             catch (Exception ex)
             {
-                return Task.FromResult(
-                    new HealthCheckResult(context.Registration.FailureStatus, exception: ex));
-            }
+				return new HealthCheckResult(context.Registration.FailureStatus, exception: ex);
+			}
         }
     }
 }
