@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Net.Http;
 
 namespace HealthChecks.UI.Configuration
 {
@@ -9,6 +11,8 @@ namespace HealthChecks.UI.Configuration
         internal int EvaluationTimeInSeconds { get; set; } = 10;
         internal int MinimumSecondsBetweenFailureNotifications { get; set; } = 60 * 10;
         internal string HealthCheckDatabaseConnectionString { get; set; }
+        internal Func<IServiceProvider, HttpMessageHandler> ApiEndpointHttpHandler { get; private set; }
+        internal Func<IServiceProvider, HttpMessageHandler> WebHooksEndpointHttpHandler { get; private set; }
 
         public Settings AddHealthCheckEndpoint(string name, string uri)
         {
@@ -48,6 +52,18 @@ namespace HealthChecks.UI.Configuration
         public Settings SetHealthCheckDatabaseConnectionString(string connectionString)
         {
             HealthCheckDatabaseConnectionString = connectionString;
+            return this;
+        }
+
+        public Settings UseApiEndpointHttpMessageHandler(Func<IServiceProvider, HttpClientHandler> apiEndpointHttpHandler)
+        {
+            ApiEndpointHttpHandler = apiEndpointHttpHandler;
+            return this;
+        }
+        
+        public Settings UseWebhookEndpointHttpMessageHandler(Func<IServiceProvider, HttpClientHandler> webhookEndpointHttpHandler)
+        {
+            WebHooksEndpointHttpHandler = webhookEndpointHttpHandler;
             return this;
         }
     }
