@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using System;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace HealthChecks.UI.Core.Data.Configuration
@@ -9,7 +10,8 @@ namespace HealthChecks.UI.Core.Data.Configuration
         public void Configure(EntityTypeBuilder<HealthCheckExecutionHistory> builder)
         {
             builder.Property(le => le.On)
-                .IsRequired(true);
+                .IsRequired(true)
+                .HasConversion(v => v, v => DateTime.SpecifyKind(v, DateTimeKind.Utc));
 
             builder.Property(le => le.Status)
                 .HasMaxLength(50)
