@@ -1,7 +1,9 @@
 ﻿using HealthChecks.UI.Configuration;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace HealthChecks.UI.Core
 {
@@ -12,21 +14,52 @@ namespace HealthChecks.UI.Core
             var resource = resources
                 .FirstOrDefault(r => r.ContentType == ContentType.HTML && r.FileName == Keys.HEALTHCHECKSUI_MAIN_UI_RESOURCE);
 
-            resource.Content = resource.Content
-                .Replace(Keys.HEALTHCHECKSUI_MAIN_UI_API_TARGET, options.ApiPath);
+            var apiPath = options.UseRelativeApiPath ? options.ApiPath.AsRelativeResource() : options.ApiPath;
 
             resource.Content = resource.Content
-                .Replace(Keys.HEALTHCHECKSUI_WEBHOOKS_API_TARGET, options.WebhookPath);
+                .Replace(Keys.HEALTHCHECKSUI_MAIN_UI_API_TARGET, apiPath);
+
+
+            var webhooksPath = options.UseRelativeWebhookPath ? options.WebhookPath.AsRelativeResource() : options.WebhookPath;
+
+            resource.Content = resource.Content
+                .Replace(Keys.HEALTHCHECKSUI_WEBHOOKS_API_TARGET, webhooksPath);
 
             
-            var isMultisegmentUI = options.UIPath.Count(c => c == '/') > 1;
-
-            var resourcePath = isMultisegmentUI ? options.ResourcesPath : options.ResourcesPath.AsRelativeResource();
+            var resourcePath = options.UseRelativeResourcesPath ? options.ResourcesPath.AsRelativeResource() : options.ResourcesPath;
             
             resource.Content = resource.Content
                 .Replace(Keys.HEALTHCHECKSUI_RESOURCES_TARGET, resourcePath);
 
             return resource;
         }
-    }
+
+        public static ICollection<UIStylesheet> GetCustomStylesheets(this UIResource resource, Options options)
+        {
+            List<UIStylesheet> styleSheets = new List<UIStylesheet>();
+            
+            if (!options.CustomStylesheets.Any())
+            {
+                resource.Content = resource.Content.Replace(Keys.HEALTHCHECKSUI_STYLESHEETS_TARGET, string.Empty);
+                return styleSheets;
+            }
+
+            foreach (var stylesheet in options.CustomStylesheets)
+            {
+                styleSheets.Add(UIStylesheet.Create(options, stylesheet));
+            }
+            
+            var htmlStyles = styleSheets.Select
+                (s =>
+            {
+                var linkHref = options.UseRelativeResourcesPath ? s.ResourcePath.AsRelativeResource() : s.ResourcePath;
+                return $"<link rel='stylesheet' href='{linkHref}'/>";
+            });
+            
+            resource.Content = resource.Content.Replace(Keys.HEALTHCHECKSUI_STYLESHEETS_TARGET,
+                string.Join("\n", htmlStyles));
+
+            return styleSheets;
+        }
+    }    
 }
