@@ -1,56 +1,78 @@
-import React from "react";
-import { Route, Link } from "react-router-dom";
-import { LivenessPage } from "./components/LivenessPage";
-import { WebhooksPage } from "./components/WebhooksPage";
-import { scaleRotate as Menu } from "react-burger-menu";
+import React from 'react';
+import { Route, Redirect, NavLink } from 'react-router-dom';
+import { LivenessPage } from './components/LivenessPage';
+import { WebhooksPage } from './components/WebhooksPage';
+import { AsideMenu } from './components/AsideMenu';
 
 interface AppProps {
-    mountPath: string;
-    apiEndpoint: string;
-    webhookEndpoint: string;
+  apiEndpoint: string;
+  webhookEndpoint: string;
 }
 
 interface AppState {
-    menuOpen: boolean;
+  menuOpen: boolean;
 }
 
-const WhiteGearIcon = require("../assets/svg/white-gear.svg");
-const WhiteHeartIcon = require("../assets/svg/white-heart.svg");
+const WhiteGearIcon = require('../assets/svg/white-gear.svg');
+const WhiteHeartIcon = require('../assets/svg/heart-check.svg');
+const SelectedHeartIcon = require('../assets/svg/heart-check.svg');
 
 export class App extends React.Component<AppProps, AppState> {
-    constructor(props: AppProps) {
-        super(props);
-        this.state = {
-            menuOpen: false
-        };
+  constructor(props: AppProps) {
+    super(props);
+    this.state = {
+      menuOpen: true
+    };
 
-        this.toggleMenu = this.toggleMenu.bind(this);
-    }
+    this.toggleMenu = this.toggleMenu.bind(this);
+  }
 
-    toggleMenu() {
-        this.setState({
-            menuOpen: !this.state.menuOpen
-        });
-    }
-    render() {
-        return <React.Fragment>
-            <div id="outer-container" style={{ height: '100%' }}>
-                <Menu onStateChange={(state) => this.setState({ menuOpen: state.isOpen })}
-                    isOpen={this.state.menuOpen}
-                    pageWrapId={'wrapper'}
-                    outerContainerId={"outer-container"} >
-                    <Link to={this.props.mountPath} className="menu-item" onClick={this.toggleMenu}>
-                        <img className="menu-icon" src={WhiteHeartIcon} />
-                        <div>Health Checks</div>
-                    </Link>
-                    <Link to={`${this.props.mountPath}/webhooks`} className="menu-item" onClick={this.toggleMenu}>
-                        <img className="menu-icon" src={WhiteGearIcon} />
-                        <div>Webhooks</div>
-                    </Link>
-                </Menu>
-                <Route exact path={this.props.mountPath} render={() => <LivenessPage endpoint={this.props.apiEndpoint} />} />
-                <Route path={`${this.props.mountPath}/webhooks`} render={() => <WebhooksPage endpoint={this.props.webhookEndpoint} />} />
-            </div>            
-        </React.Fragment>
-    }
+  toggleMenu() {
+    this.setState({
+      menuOpen: !this.state.menuOpen
+    });
+  }
+  render() {
+    return (
+      <main id="outer-container">
+        <AsideMenu
+          isOpen={this.state.menuOpen}
+          onClick={() => this.setState({ menuOpen: !this.state.menuOpen })}>
+          <NavLink
+            to="/healthchecks"
+            className="hc-aside-menu__item"
+            activeClassName="hc-aside-menu__item--active"
+            >
+            <img alt="icon heart check" className="hc-menu-icon" src={WhiteHeartIcon} />
+            <span>Health Checks</span>
+          </NavLink>
+          <NavLink
+            to="/webhooks"
+            className="hc-aside-menu__item"
+            activeClassName="hc-aside-menu__item--active"
+            >
+            <img alt="icon gear" className="hc-menu-icon" src={WhiteGearIcon} />
+            <span>Webhooks</span>
+          </NavLink>
+        </AsideMenu>
+        <section className="hc-section-router">
+          <Route
+            exact
+            path="/"
+            render={() => <Redirect to="/healthchecks" />}
+          />
+          <Route
+            path="/healthchecks"
+            render={() => <LivenessPage endpoint={this.props.apiEndpoint} />}
+          />
+          <Route
+            path="/webhooks"
+            render={() => (
+              <WebhooksPage endpoint={this.props.webhookEndpoint} />
+            )}
+          />
+        </section>
+      </main>
+    );
+  }
 }
