@@ -1,23 +1,21 @@
-export const
-    statusUp: string = "Up",
-    statusDown: string = "Down",
-    statusDegraded: string = "Degraded";
+export const statusUp: string = 'Up',
+  statusDown: string = 'Down',
+  statusDegraded: string = 'Degraded';
 
-const okImage = require("../assets/images/ok.png");
-const downImage = require("../assets/images/down.png");
-const degradedImage = require("../assets/images/degraded.png");
 const kubernetesIcon = require('../assets/images/kubernetes-icon.png');
 
 const imageResources = [
-    { state: 'Failed', image: downImage },
-    { state: 'Unhealthy', image: downImage },
-    { state: 'Degraded', image: degradedImage },
-    { state: 'Healthy', image: okImage }
-]
-
-export const discoveryServices = [
-    { name: 'kubernetes', image: kubernetesIcon }
+  { state: 'Failed', image: 'error', color: '--dangerColor' },
+  { state: 'Unhealthy', image: 'error', color: '--dangerColor' },
+  { state: 'Degraded', image: 'warning', color: '--warningColor' },
+  { state: 'Healthy', image: 'check_circle', color: '--successColor' }
 ];
 
-const getStatusImage = (status: string) => imageResources.find(s => s.state == status)!.image;
-export { getStatusImage };
+export const discoveryServices = [
+  { name: 'kubernetes', image: kubernetesIcon }
+];
+
+const getStatusConfig = (status: string) =>
+  imageResources.find(s => s.state == status);
+
+export { getStatusConfig };
