@@ -25,7 +25,6 @@ namespace UnitTests.HealthChecks.DependencyInjection.SqlServer
 
             registration.Name.Should().Be("sqlserver");
             check.GetType().Should().Be(typeof(SqlServerHealthCheck));
-
         }
 
         [Fact]
@@ -43,6 +42,29 @@ namespace UnitTests.HealthChecks.DependencyInjection.SqlServer
 
             registration.Name.Should().Be("my-sql-server-1");
             check.GetType().Should().Be(typeof(SqlServerHealthCheck));
+        }
+
+        [Fact]
+        public void add_health_check_with_connection_string_factory_when_properly_configured()
+        {
+            var services = new ServiceCollection();
+            var factoryCalled = false;
+            services.AddHealthChecks()
+                .AddSqlServer(_ =>
+                {
+                    factoryCalled = true;
+                    return "connectionstring";
+                });
+
+            var serviceProvider = services.BuildServiceProvider();
+            var options = serviceProvider.GetService<IOptions<HealthCheckServiceOptions>>();
+
+            var registration = options.Value.Registrations.First();
+            var check = registration.Factory(serviceProvider);
+
+            registration.Name.Should().Be("sqlserver");
+            check.GetType().Should().Be(typeof(SqlServerHealthCheck));
+            factoryCalled.Should().BeTrue();
         }
     }
 }
